+++ webpack.config.js
@@ -6,10 +6,10 @@ const path = require('path');
 const config = {
     target: 'node',
 
-    entry: './extension.js',
+    entry: './src/extension.js',
     output: {
         path: path.resolve(__dirname, 'dist'),
-        filename: 'extension.js',
+        filename: 'src/extension.js',
         libraryTarget: "commonjs2",
         devtoolModuleFilenameTemplate: "../[resource-path]",
     },