+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 
-All notable changes to the "vscode-disco" extension will be documented in this file.
+## [0.0.5] - 2019-09-19
+- Modular code
+- Commands renamed
+- Added two commands - Disco Increase Interval by 100 ms and Disco Decrease Interval by 100 ms
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.0.4] - 2019-09-19
+- Made compatible with vs code > 1.0.0
+- Added icon
+- Changes in manifest
 
 ## [0.0.3] - 2019-09-19
 ### Added